refactor: Move main/test handler
